% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal_tree_sampling_aux.R
\name{ess_value}
\alias{ess_value}
\title{Estimated Sample Size}
\usage{
ess_value(weights, log)
}
\arguments{
\item{weights}{a vector of weight values associated to the sampled trajectories}

\item{log}{boolean, \code{TRUE} if log weights are provided}
}
\value{
a number
}
\description{
Returns the estimated number of independent samples based on the weights, using Rényi's entropy: \eqn{\frac{1}{\sum_i w_i^2}}.
}
\examples{
ess_value(weights=c(3,0,456,1,0,0,5,57), log=FALSE)

}
