% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal_tree_sampling_aux.R
\name{height_distribution}
\alias{height_distribution}
\title{Compute the Height Distribution}
\usage{
height_distribution(trees, t_max, print = FALSE, plot = FALSE,
  save = FALSE, filename = "height_dist.RData")
}
\arguments{
\item{trees}{a list of \code{tree} objects}

\item{t_max}{time of the last sampling event}

\item{print}{boolean, if \code{TRUE} prints the median, mean and 95\% HPD interval for the distribution}

\item{plot}{boolean, if \code{TRUE} plot the histogram of the distribution}

\item{save}{boolean, if \code{TRUE} save the height distribution in a .RData file}

\item{filename}{if \code{save} is \code{TRUE}, name of the file who will contain the distribution}
}
\value{
a vector containing the heights of the trees specified in \code{trees}
}
\description{
Compute the height distribution for a list of trees. The height of a tree is the time elapsed between the most recent common ancestor of the samples and the last sample.
}
\examples{
height_distribution(trees=list(list('branch_length'=1,
subtrees=list(list('branch_length'=1, subtrees=NULL)))), t_max=2, print=TRUE, plot=TRUE)

}
