% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal_tree_sampling_aux.R
\name{newick}
\alias{newick}
\title{Save Trees in Newick Format}
\usage{
newick(trees, filename)
}
\arguments{
\item{trees}{list of \code{tree} objects}

\item{filename}{name of the file to save the trees}
}
\description{
Write a list of trees in Newick format in the specified text file.
}
\examples{
newick(trees=list(list('branch_length'=1, subtrees=list(list('branch_length'=1, subtrees=NULL)))),
filename='one_tree.txt')

}
