% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal_tree_sampling_aux.R
\name{sample_trees}
\alias{sample_trees}
\title{Tree Sampling}
\usage{
sample_trees(k1 = NULL, e = NULL, k2, sampling_times, lambda, mu, psi,
  r)
}
\arguments{
\item{k1}{number of trajectories to simulate, from which \code{k2} trajectories will be subsampled}

\item{e}{ESS value to achieve}

\item{k2}{number of trees to simulate}

\item{sampling_times}{vector containing the sampling times \eqn{(s_j)_{j=1...k}} in increasing order}

\item{lambda}{birth rate}

\item{mu}{death rate}

\item{psi}{sampling rate}

\item{r}{removal probability}
}
\value{
a list of \code{tree} objects
}
\description{
Samples trees under a birth death model conditioned on sampling times.
\itemize{
    \item If \code{k1} is specified then the function simulates \code{k1} weighted trajectories, then subsample \code{k2} trajectories from them.
    \item If \code{e} is specified the function will simulate just enough weighted trajectories to obtain an ESS value greater or equal to \code{e}, and then subsample \code{k2} trajectories from them.
}
In both cases, the function then reconstructs one tree per trajectory and returns these trees.
One of \code{k1} or \code{e} has to be specified. If both are specified, \code{e} will be ignored.
}
\examples{
sample_trees(k1=NULL, e=100, k2=100, sampling_times=c(1,2), lambda=2, mu=1, psi=0.3, r=0.2)

}
