% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal_tree_sampling_aux.R
\name{str_tree}
\alias{str_tree}
\title{Write Tree in Newick Format}
\usage{
str_tree(tree)
}
\arguments{
\item{tree}{an object of type \code{tree}}
}
\value{
a string
}
\description{
Write the Newick description of a tree, and return the corresponding string.
}
\examples{
str_tree(tree=list('branch_length'=1, 'subtrees'=list(list('branch_length'=1, 'subtrees'=NULL))))

}
