% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal_tree_sampling_aux.R
\name{trajectory}
\alias{trajectory}
\title{Trajectory Simulation}
\usage{
trajectory(lambda, mu, r, sampling_times)
}
\arguments{
\item{lambda}{birth rate}

\item{mu}{death rate}

\item{r}{removal probability, the probability for an individual to be removed upon sampling}

\item{sampling_times}{vector containing the sampling times \eqn{(s_j)_{j=1...k}} in increasing order}
}
\value{
an object of the class \code{trajectory}, containing:
\itemize{
   \item \code{t} a sequence of time points \eqn{(t_i)_{i=1...n}} with \eqn{t_1=0} and \eqn{t_n=s_k},
   \item \code{N} a sequence of population size values where \eqn{N_i} is the value on the interval \eqn{[t_i,t_{i+1})}, and
   \item \code{removed} a vector with the sampling times of removed samples.
}
}
\description{
Simulates a trajectory of the population size under a constant-rate birth-death model as well as the removal vector \eqn{y} (\eqn{y_i=1} iff \eqn{s_i} is removed upon sampling). It uses Gillespie algorithm to simulate portions of trajectory in the time intervals between removals.
}
\examples{
trajectory(lambda=2, mu=1, r=0.2, sampling_times=c(1,2))

}
