% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal_tree_sampling_aux.R
\name{tree_from_trajectory}
\alias{tree_from_trajectory}
\title{Tree Simulation from a Trajectory}
\usage{
tree_from_trajectory(traj, sampling_times)
}
\arguments{
\item{traj}{an object of class \code{trajectory}}

\item{sampling_times}{vector containing the sampling times \eqn{(s_j)_{j=1...k}} in increasing order}
}
\value{
an object of class \code{tree}, containing an incident branch length \code{branch_length} and a list of 0 to 2 subtrees \code{subtrees}
}
\description{
Reconstructs a tree from a trajectory and known sampling times.
}
\examples{
tree_from_trajectory(traj=list('t'=c(0,0.5,1,2), 'N'=c(1,2,1,1),
'removed'=c(1)), sampling_times=c(1,2))

}
