% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal_tree_sampling_aux.R
\name{tree_height}
\alias{tree_height}
\title{Tree Height}
\usage{
tree_height(tree, t_max)
}
\arguments{
\item{tree}{an object of class \code{tree}}

\item{t_max}{the last sampling time}
}
\value{
a number
}
\description{
Compute the height of a tree, that is the time elapsed between the most recent common ancestor of the samples and the last sample.
}
\examples{
tree_height(tree=list('branch_length'=1, 'subtrees'=list(list('branch_length'=1,
'subtrees'=NULL))), t_max=2)

}
