% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal_tree_sampling_aux.R
\name{weight}
\alias{weight}
\title{Weight of a Trajectory}
\usage{
weight(traj, sampling_times, psi)
}
\arguments{
\item{traj}{object of class \code{trajectory}}

\item{sampling_times}{vector containing the sampling times \eqn{(s_j)_{j=1...k}} in increasing order}

\item{psi}{sampling rate}
}
\value{
a number
}
\description{
Computes the log probability density of the sampling times given the trajectory.
}
\examples{
weight(traj=list('t'=c(0,0.5,1,2), 'N'=c(1,2,1,1), 'removed'=c(1)),
sampling_times=c(1,2), psi=0.3)

}
